Fixes shutter-project/shutter#274 - continue numbering even if bigger than 999 (n_counter template value)
In the case where an autoincrement is used with a filename template like
"...%NNN", continue the search for the next increment value after the
maximum of the template (ex.: 999).

Little behavior change:
The increment counter is specific to each parsed filename template (ex.
Action).
So for example, you can have:
- Selection_013.jpg
- Workspace_135.jpg
- Selection_014.jpg
This is needed to be sure to avoid false positives with unrelated files
like:
File in folder: Selection_160_fvcdf_Balbla_17062020.jpg
next counter = 17062021

For that behavior, the fct_get_next_filename function call was moved
after the screenshot is taken to that we already have access to
screenshot and screenshooter.
Also, the shutter wildcard parsing is now done before the call to
fct_get_next_filename.

All in all, it should even make the code a little bit faster.
Having the filename/counter search performed after the screenshot also
improves the perceived reactivity of the application as the screenshot
starts directly after a click on a button, when it is the most needed in
most cases.

In addition, a little bit of cleanup/simplification was done on this
part of the code.